requirements.txt
Django==6.0.4
djangorestframework==3.14.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.2
google-auth==2.28.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
